Move StatementUtils to presto-analyzer
